Ignore targets missing from graph
Depending on the query that was used to create a `DependenciesIndex`,
targets can have dependencies to targets that are not defined in the
query results.

For instance, the query executed by the lint command reads:

    allpaths(foo, @maven//:all)

If `foo` has a dependency on a target `bar`, which has no dependency on
any of `@maven//:all`, then `bar` will not be defined in the results of
the query. Getting the `dependencies` of `foo` would then cause a crash
in multiversion.

This commit changes `DependenciesIndex` so that the missing targets are
simply ignored. Because these targets do not appear on any path ending
in any of `@maven//:all`, then they cannot introduce conflicts, and
ignoring them is safe.