Emit YAML linting report
The lint command now emits a YAML linting report at the location
specified by `--lint-report-path` when the option is set.

The option can also be set from `export` and `pants-export`.

The report contains the roots that failed and the name and versions of
the conflicting dependencies:

```
"//my:target":
  - "com.acme:other-dependency": ["1.1", "2.0"]
  - "my.org:my-dependency": ["1.0", "1.1"]
"//my/other:target":
  - "org:dep": ["0.1", "0.2", "1.0"]
```