Infer and de-duplicate dependency transformations
At the moment, the inferred transformations are ignored and never used.

The `exclusion` and `dependencies` of the various third party
dependencies that are defined in the ThirdpartyConfig can be seen as
defining global and local transformations of the dependency graph. There
are 3 kinds of transformations that are defined:

 - Addition: This transformation adds a dependency from one module to
   another module. It is created by adding a `dependency` without a
   matching exclusion. When the dependency is resolved, the added
   dependencies will be resolved together.
 - Removal: This transformations excludes prevents a given dependency
   from being resolved. It is created by adding an `exclusion` clause
   that has no matching `dependency`.
 - Replacement: This transformation forces a specific version of a
   dependency. It is created by adding a `dependency` with a matching
   `exclusion`.

All these transformations can be understood as either global or local.
Global transformations are defined on canonical definitions (the
definitions whose target package matches the organization name). All
other transformations are local.  Global transformations apply to all
dependency resolutions, whereas local transformations apply only to the
targets where they're defined and will introduce a fork in the
dependency resolution graph.