Apply global exclusions only to matching modules
Previously, global exclusion transformations would prevent excluded
modules from appearing in any resolution's results. Instead, this commit
changes the application of the global exclusion transformations so that they
apply only on the resolutions that include the module on which the
transformation is defined.

For instance, if we consider the following third party configuration:

```
 - com.google.guava:guava:27.1-jre,
     excludes org.checkerframework:checker-qual
     (defining a global exclusion)
 - org.apache.pulsar:pulsar-client:2.4.2,
     depends on org.checkerframework:checker-qual:2.0.0
 - com.google.inject:guice:4.2.4,
     depends on com.google.guava:guava:4.2.3
```

Then, the resolution of guava will not include
`org.checkerframework:checker-qual` (because it is excluded). The
resolution of `com.google.inject:guice` will not include it either,
because `guice` transitively depends on `guava` which excludes
`checker-qual`.  Finally, `pulsar-client` will include the dependency on
`checker-qual`.  Prior to this commit, `pulsar-client` wouldn't include
the dependency on `checker-qual`, because it would be globally excluded.