Empty resolutions still create targets
Previously, we would not create a target matching the third party root
(ie. `@maven//:3rdparty/jvm/foo_bar`) when the resolution yield only
artifacts that appear to be unrelated to the resolved root dependency.
This situation arises, for instance, when resolving a classifier that
does not exist.

This commit ensures that we still generate the third party root target.
In the test introduced by this commit, the target
`@maven//:with-classifier` wouldn't exist prior to this patch.