Print resolution info in Bazel file
Previously, the Bazel file produced by multiversion would not include
information about the context of the resolution that produced a given
target. This commit adds a comment to every declared dependency, so that
one can get information about the originating target, the exclusions and
the dependencies that affected the resolution.