Add support for round-trip dependencies
Round-trip dependencies are dependencies that are published, but also
exist as source in the repository. Because these are published, they can
appear as dependencies of other resolved artifacts. However, these
artifacts shouldn't be resolved and a dependency should be added on the
source target instead.

This commit implements support for round-trip dependencies by leveraging
the `overriddeTargets` attribute in the input export. This attribute
contains a list of mapping from modules to the source target that they
should replace.

After the first resolution, the resolution configuration that depend on
one overridden dependencies are re-configured so that these dependencies
are excluded, and the source dependency is added instead.