Avoid missing targets in generated graph
Previously, the generated third party dependency graph was sometimes
missing some of the targets representing the transitive resolution of a
dependency.

The faulty logic that decided whether such a target needs to be emitted
relied on matching the resolved artifact with the resolution that pulled
the artifact in (they had to have matching organization, name and
version). This is flawed, because we have no guarantee that the artifact
that is eventually selected will necessarily be coming from the
resolution of that module in particular.

This logic is rewritten to rely solely on the resolution index, which is
now correctly populated with the reconciled versions.