Optionally write resolution manifests
The resolution manifests include information about all the transitive
dependencies that are included as part of the resolution for every third
party target.

The resolution manifests are written if and only if `--manifests-root`
is passed to multiversion. This path is used as the root under which the
manifests should be written. One manifest is written per resolved
target.

Given a manifest root set to `foo/bar`, and a target `fizz/buzz:quux`,
the manifest will be written in `foo/bar/fizz/buzz/bazel-deps-quux.txt`.