Make direct forced win over transitives
Problem
-------
Currently the `force=True` on a dependency carries over to the
transitive dependencies.
This allows transitive dependencies to unintentionally evict the direct
dependencies.
In a monorepo, this is a surprising behavior.

Solution
--------
Don't carry over the `force=True` over to the transitives.
This would mean that silently the transitive dependencies would be
downgraded.