Show version origin in declared eviction linter
Previously, when a declared dependency was evicted in favor of a
different version transitively pulled in by another dependency, we
wouldn't give the user an easy way to find what dependency caused the
new version to be pulled in.

This commit improves the linter so that it shows where the different
versions come from.