Support setting the URL of the JAR to resolve
Previously, multiversion didn't have an option to manually force the URL
of a JAR to resolve, which made it harder for developers to test targets
with a custom built JAR.

This commit adds an attribute `url` to the `DependencyConfig` where the
URL of the JAR to resolve can be set. Whether multiversion should allow
`url` in the dependency configuration can be controller with
`--[no-]allow-url`.

Following the Coursier semantics of the `url=...` attribute,
dependencies that set the URL will be resolve intransitively.
Dependencies that set the URL but are not marked as intransitive will
trigger a warning message, and the resolution will be done
intransitively.

Dependencies that set the URL, but are evicted will trigger an error,
regardless of `--no-fail-on-evicted-declared`.