Fix missing targets when deps are duplicated
Previously, if the same third party dependency was defined by several
targets (before or after reconciliation), then multiversion generated a
graph that was missing some nodes.

The root cause is that when dependencies were de-duplicated, the
different targets were not taken into account, and some targets were
then associated with no resolution.

This commit fixes it by taking the targets for which the resolution is
performed into account when de-duplicating the dependencies.