Resolve selected transitive dependencies
Previously, when a transitive dependency D was evicted in favor of
another, and the transitive dependencies of D changed between the 2
versions, we would end up with incorrect transitive dependencies.

To fix this, we re-run the resolution of all declared dependencies along
with their transitive dependencies, when they're affected by eviction.